requirements.txt
transformers
datasets
tiktoken
litellm
jsonlines
python-dateutil
anthropic
google-genai
